+++ nest-forms-backend/src/nases/nases.service.ts
@@ -33,6 +33,7 @@ import {
 } from '../files/files.errors.enum'
 import FilesService from '../files/files.service'
 import { RabbitPayloadDto } from '../form-delivery-consumer/dtos/form-delivery-consumer.dto'
+import { SendFormResponseDto } from '../form-sender/dtos/responses.dto'
 import FormValidatorRegistryService from '../form-validator-registry/form-validator-registry.service'
 import { FormUpdateBodyDto } from '../forms/dtos/requests.dto'
 import {
@@ -44,7 +45,6 @@ import RabbitmqClientService from '../rabbitmq-client/rabbitmq-client.service'
 import { ErrorsEnum } from '../utils/global-enums/errors.enum'
 import ThrowerErrorGuard from '../utils/guards/thrower-error.guard'
 import { LineLoggerSubservice } from '../utils/subservices/line-logger.subservice'
-import { SendFormResponseDto } from './dtos/responses.dto'
 import { verifyFormSignatureErrorMapping } from './nases.errors.dto'
 import { NasesErrorsEnum, NasesErrorsResponseEnum } from './nases.errors.enum'
 import NasesSenderService from './services/nases.sender.service'
